Enhancement: test updates
